refactor: SectionQueryResponse inner class에서 별도로 분리
